requirements.txt
# MCP EDA Server — Dependencies
mcp[cli]
uvicorn
python-dotenv
